+++ lib/Service/BinExt.php
@@ -9,7 +9,7 @@ use OCA\Memories\Settings\SystemConfig;
 class BinExt
 {
     public const EXIFTOOL_VER = '12.60';
-    public const GOVOD_VER = '0.1.20';
+    public const GOVOD_VER = '0.1.21';
     public const NX_VER_MIN = '1.1';
 
     /** Get the path to the temp directory */
@@ -176,6 +176,8 @@ class BinExt
     {
         // Get config from system values
         $env = [
+            'qf' => SystemConfig::get('memories.vod.qf'),
+
             'vaapi' => SystemConfig::get('memories.vod.vaapi'),
             'vaapiLowPower' => SystemConfig::get('memories.vod.vaapi.low_power'),