+++ lib/Settings/SystemConfig.php
@@ -83,6 +83,10 @@ class SystemConfig
         // start go-vod if it is not running already.
         'memories.vod.external' => false,
 
+        // Quality Factor used for transcoding
+        // This correspondes to CRF for x264 and global_quality for VA-API
+        'memories.vod.qf' => 24,
+
         // Set the default video quality for a first time user
         //    0 => Auto (default)
         //   -1 => Original (max quality with transcoding)